+++ app/models/sheet_transaction.rb
@@ -17,7 +17,7 @@ class SheetTransaction < ActiveRecord::Base
       sheet.update(sheet_params)
     end
 
-    ignore_attributes = %w(created_at updated_at authentication_token deleted last_viewed_at)
+    ignore_attributes = %w(created_at updated_at authentication_token deleted)
 
     original_attributes = sheet.previous_changes.collect{|k,v| [k,v[0]]}.reject{|k,v| ignore_attributes.include?(k.to_s)}