+++ db/migrate/20150430133644_remove_last_viewed_at_and_last_viewed_by_id_from_sheets.rb
@@ -1,0 +1,16 @@
+class RemoveLastViewedAtAndLastViewedByIdFromSheets < ActiveRecord::Migration
+  def up
+    remove_index :sheets, :last_viewed_by_id
+
+    remove_column :sheets, :last_viewed_by_id, :integer
+    remove_column :sheets, :last_viewed_at, :datetime
+  end
+
+  def down
+    add_column :sheets, :last_viewed_by_id, :integer
+    add_column :sheets, :last_viewed_at, :datetime
+
+    add_index :sheets, :last_viewed_by_id
+  end
+
+end